Lab 4, 5 is added

+++ src/LAB2/Calculator.java
@@ -39,14 +39,14 @@ public class Calculator {
     }
     public static double div(int a, int b)
     {
-        return a/b;
+        return (double)a/b;
     }
     public static double div (double a,double b)
     {
-        return a/b;
+        return  a/b;
     }
     public static double div(long a,long b)
     {
-        return a/b;
+        return (double)a/b;
     }
 }
